+++ backend/canisters/group/c2c_client/src/lib.rs
@@ -8,9 +8,9 @@ generate_c2c_call!(c2c_events_by_index);
 generate_c2c_call!(c2c_events_internal);
 generate_c2c_call!(c2c_events_window);
 generate_c2c_call!(c2c_name_and_members);
-generate_c2c_call!(c2c_summary);
-generate_c2c_call!(c2c_summary_updates);
 generate_c2c_call!(public_summary);
+generate_c2c_call!(summary);
+generate_c2c_call!(summary_updates);
 
 generate_candid_c2c_call!(events);
 generate_candid_c2c_call!(invite_code);